requirements.txt
fastapi==0.115.14
numpy>=2.3.1
nltk==3.9.1
pydantic~=2.11.7
pytest==8.4.1
sentence_transformers==5.0.0
uvicorn==0.35.0
pandas==2.3.1
pandas-stubs==2.3.0.250703
duckdb==1.3.2

# TODO: Look which of these are actually needed
# Retrieval engine dependencies
# scikit-learn>=1.4
# tqdm>=4.66
# transformers>=4.42
# torch>=2.3.0

# Development dependencies
# black>=24.3
# isort>=5.13
# mypy>=1.10
